backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.12.5
pydantic-settings==2.4.0
anthropic==0.34.0
openai>=1.40.0
chromadb==0.5.5
sentence-transformers==3.1.1
torch==2.4.1
PyMuPDF==1.24.10
python-docx==1.1.2
aiosqlite==0.20.0
loguru==0.7.2
tenacity==9.0.0
pytest==8.3.0
pytest-asyncio==0.23.0
httpx==0.27.0
